+++ R/sits_source_api.R
@@ -6,6 +6,8 @@
 #' These functions provide an API to handle/retrieve data from sources.
 #'
 #' @param source     A \code{character} value referring to a valid data source.
+#' @param internal   A \code{logical} value if internal sources should
+#' be listed. Internal sources are: 'PROBS', 'CLASSIFIED', and 'LOCAL'
 #'
 #' @return
 #' The values returned by each function are described as follows.
@@ -17,7 +19,7 @@ NULL
 #'
 #' @return \code{.sources()} returns a \code{character} vector
 #' with all sources names available in sits.
-.sources <- function() {
+.sources <- function(internal = TRUE) {
 
     res <- .config_names(key = c("sources"))
 
@@ -27,6 +29,11 @@ NULL
     # post-condition
     .check_chr(res, allow_empty = FALSE, len_min = 1,
                msg = "invalid 'sources' in config file")
+
+    # filter internal sources
+    if (!internal) {
+        res <- res[!res %in% c("PROBS", "CLASSIFIED", "LOCAL")]
+    }
 
     return(res)
 }
@@ -203,8 +210,7 @@ NULL
     res <- toupper(res)
 
     if (!add_cloud)
-        res <- res[res != .source_cloud(source = source,
-                                        collection = collection)]
+        res <- res[res != .source_cloud()]
 
     if (!is.null(fn_filter)) {
         select <- vapply(res, function(band) {
@@ -419,8 +425,7 @@ NULL
 #'
 #' @return \code{.source_cloud()} returns a \code{character} vector with cloud
 #' band name.
-.source_cloud <- function(source,
-                          collection) {
+.source_cloud <- function() {
 
     return("CLOUD")
 }
@@ -444,8 +449,7 @@ NULL
     .source_collection_check(source = source, collection = collection)
 
     res <- .config_get(key = c("sources", source, "collections", collection,
-                               "bands", .source_cloud(source = source,
-                                                      collection = collection),
+                               "bands", .source_cloud(),
                                "bit_mask"))
 
     # post-condition
@@ -475,8 +479,7 @@ NULL
     .source_collection_check(source = source, collection = collection)
 
     res <- .config_get(key = c("sources", source, "collections", collection,
-                               "bands", .source_cloud(source = source,
-                                                      collection = collection),
+                               "bands", .source_cloud(),
                                "values"))
 
     # post-condition
@@ -506,8 +509,7 @@ NULL
     .source_collection_check(source = source, collection = collection)
 
     res <- .config_get(key = c("sources", source, "collections", collection,
-                               "bands", .source_cloud(source = source,
-                                                      collection = collection),
+                               "bands", .source_cloud(),
                                "interp_values"))
 
     # post-condition
@@ -884,11 +886,15 @@ NULL
 #'
 #' @return \code{.source_items_get_sensor()} returns a \code{character} value.
 #'
-.source_items_get_sensor <- function(source, items, ..., collection = NULL) {
+.source_collection_sensor <- function(source, collection) {
 
-    s <- .source_new(source)
+    res <- .config_get(key = c("sources", source, "collections",
+                               collection, "sensor"))
 
-    UseMethod(".source_items_get_sensor", s)
+    .check_chr(res, allow_null = TRUE,
+               msg = "invalid 'sensor' value")
+
+    return(res)
 }
 
 #' @rdname source_cube
@@ -899,11 +905,15 @@ NULL
 #' @return \code{.source_items_get_satellite()} returns a \code{character}
 #' value.
 #'
-.source_items_get_satellite <- function(source, items, ..., collection = NULL) {
+.source_collection_satellite <- function(source, collection) {
 
-    s <- .source_new(source)
+    res <- .config_get(key = c("sources", source, "collections",
+                               collection, "satellite"))
 
-    UseMethod(".source_items_get_satellite", s)
+    .check_chr(res, allow_null = TRUE,
+               msg = "invalid 'satellite' value")
+
+    return(res)
 }
 
 #' @rdname source_cube