+++ R/sits_values.R
@@ -26,15 +26,18 @@
 #' ls3 <- sits_values(cerrado_2classes[1:2, ], format = "bands_cases_dates")
 #' @export
 sits_values <- function(data, bands = NULL, format = "cases_dates_bands") {
+
+    # set caller to show in errors
+    .check_set_caller("sits_values")
+
     .check_chr_within(
         x = format,
         within = c("cases_dates_bands",
                    "bands_cases_dates",
                    "bands_dates_cases"),
         discriminator = "one_of",
-        msg = paste("sits_values: valid format parameter are",
-                    "'cases_dates_bands', 'bands_cases_dates'",
-                    "or 'bands_dates_cases'")
+        msg = paste("valid format parameter are 'cases_dates_bands',",
+                    "'bands_cases_dates' or 'bands_dates_cases'")
     )
 
     class(format) <- c(format, class(format))