+++ tests/testthat/test-accuracy.R
@@ -52,8 +52,8 @@ test_that("Accuracy areas", {
     cube <- sits_cube(
         source = "LOCAL",
         name = "sinop-2014",
-        satellite = "TERRA",
-        sensor = "MODIS",
+        origin = "BDC",
+        collection = "MOD13Q1-6",
         data_dir = data_dir,
         delim = "_",
         parse_info = c("X1", "X2", "tile", "band", "date")