+++ tests/testthat/test-data.R
@@ -306,8 +306,8 @@ test_that("Reading a LAT/LONG from RASTER", {
         sits_cube(
             source = "LOCAL",
             name = "sinop-2014",
-            satellite = "TERRA",
-            sensor = "MODIS",
+            origin = "BDC",
+            collection = "MOD13Q1-6",
             data_dir = data_dir,
             delim = "_",
             parse_info = c("X1", "X2", "tile", "band", "date")
@@ -337,8 +337,8 @@ test_that("Reading a CSV file from RASTER", {
         sits_cube(
             source = "LOCAL",
             name = "sinop-2014",
-            satellite = "TERRA",
-            sensor = "MODIS",
+            origin = "BDC",
+            collection = "MOD13Q1-6",
             data_dir = data_dir,
             delim = "_",
             parse_info = c("X1", "X2", "tile", "band", "date")